Move augur.utils.load_sequences and helpers to augur.io.sequences
Main motivation is to reduce complexity of augur.utils and the chance of
cyclic imports. The chance of cyclic imports has increased after
"Replace usage of Bio.SeqIO.read with read_sequence" (6fc90ce) which
added an import of augur.io.sequences in augur.utils, effectively
preventing augur.io.sequences from importing augur.utils.

Moving all sequence-related functions to augur.io.sequences restores the
ability for augur.io.sequences to import from augur.utils.
