+++ augur/utils.py
@@ -10,7 +10,7 @@ from .__version__ import __version__
 
 from augur.data import as_file
 from augur.io.file import open_file
-from augur.io.sequences import read_single_sequence
+from augur.io.print import print_err
 
 from augur.types import ValidationMode
 from augur.errors import AugurError
@@ -165,275 +165,11 @@ class AugurJSONEncoder(json.JSONEncoder):
         return super().default(obj)
 
 
-def load_features(reference, feature_names=None):
-    """
-    Parse a GFF/GenBank reference file. See the docstrings for _read_gff and
-    _read_genbank for details.
+def load_features(*args, **kwargs):
+    print_err("WARNING: augur.utils.load_features has been moved to augur.io.sequences.load_features.")
+    from augur.io.sequences import load_features as new_location
+    return new_location(*args, **kwargs)
 
-    Parameters
-    ----------
-    reference : str
-        File path to GFF or GenBank (.gb) reference
-    feature_names : None or set or list (optional)
-        Restrict the genes we read to those in the set/list
-
-    Returns
-    -------
-    features : dict
-        keys: feature names, values: :py:class:`Bio.SeqFeature.SeqFeature` Note
-        that feature names may not equivalent to GenBank feature keys
-
-    Raises
-    ------
-    AugurError
-        If the reference file doesn't exist, or is malformed / empty
-    """
-    #checks explicitly for GFF otherwise assumes Genbank
-    if not os.path.isfile(reference):
-        raise AugurError(f"reference sequence file {reference!r} not found")
-
-    if '.gff' in reference.lower():
-        return _read_gff(reference, feature_names)
-    else:
-        return _read_genbank(reference, feature_names)
-
-def _read_nuc_annotation_from_gff(record, reference):
-    """
-    Looks for the ##sequence-region pragma as well as 'region' & 'source' GFF
-    types. Note that 'source' isn't really a GFF feature type, but is used
-    widely in the Nextstrain ecosystem. If there are multiple we check that the
-    coordinates agree.
-
-    Parameters
-    ----------
-    record : :py:class:`Bio.SeqRecord.SeqRecord`
-    reference: string
-        File path to GFF reference
-
-    Returns
-    -------
-    :py:class:`Bio.SeqFeature.SeqFeature`
-
-    Raises
-    ------
-    AugurError
-        If no information on the genome / seqid length is available or if the
-        information is contradictory
-    """
-    nuc = {}
-    # Attempt to parse the sequence-region pragma to learn the genome
-    # length (in the absence of record/source we'll use this for 'nuc')
-    sequence_regions = record.annotations.get('sequence-region', [])
-    if len(sequence_regions)>1:
-        raise AugurError(f"Reference {reference!r} contains multiple ##sequence-region pragma lines. Augur can only handle GFF files with a single one.")
-    elif sequence_regions:
-        from Bio.SeqFeature import SeqFeature, FeatureLocation
-        (name, start, stop) = sequence_regions[0]
-        nuc['pragma'] = SeqFeature(
-            FeatureLocation(start, stop, strand=1),
-            type='##sequence-region pragma',
-            id=name,
-        )
-    for feat in record.features:
-        if feat.type == "region":
-            nuc['region'] = feat
-        elif feat.type == "source":
-            nuc['source'] = feat
-
-    # ensure they all agree on coordinates, if there are multiple
-    if len(nuc.values())>1:
-        coords = [(name, int(feat.location.start), int(feat.location.end)) for name,feat in nuc.items()]
-        if not all(el[1]==coords[0][1] and el[2]==coords[0][2] for el in coords):
-            raise AugurError(f"Reference {reference!r} contained contradictory coordinates for the seqid/genome. We parsed the following coordinates: " +
-                             ', '.join([f"{el[0]}: [{el[1]+1}, {el[2]}]" for el in coords]) # +1 on the first coord to shift to one-based GFF representation
-                             )
-
-    if 'pragma' in nuc: ## the pragma is GFF's preferred way to define nuc coords
-        return nuc['pragma']
-    elif 'region' in nuc:
-        return nuc['region']
-    elif 'source' in nuc:
-        return nuc['source']
-    else:
-        raise AugurError(f"Reference {reference!r} didn't define any information we can use to create the 'nuc' annotation. You can use a line with a 'record' or 'source' GFF type or a ##sequence-region pragma.")
-
-
-def _read_gff(reference, feature_names):
-    """
-    Read a GFF file. We only read GFF IDs 'gene' or 'source' (the latter may not technically
-    be a valid GFF field, but is used widely within the Nextstrain ecosystem).
-    Only the first entry in the GFF file is parsed.
-    We create a "feature name" via:
-    - for 'source' IDs use 'nuc'
-    - for 'gene' IDs use the 'gene', 'gene_name' or 'locus_tag'.
-      If none are specified, the intention is to silently ignore but there are bugs here.
-
-    Parameters
-    ----------
-    reference : string
-        File path to GFF reference
-    feature_names : None or set or list
-        Restrict the genes we read to those in the set/list
-
-    Returns
-    -------
-    features : dict
-        keys: feature names, values: :py:class:`Bio.SeqFeature.SeqFeature`
-        Note that feature names may not equivalent to GenBank feature keys
-
-    Raises
-    ------
-    AugurError
-        If the reference file contains no IDs or multiple different seqids
-        If a gene is found with the name 'nuc'
-    """
-    from BCBio import GFF
-    valid_types = ['gene', 'source', 'region']
-    features = {}
-
-    with open_file(reference) as in_handle:
-        # Note that `GFF.parse` doesn't always yield GFF records in the order
-        # one may expect, but since we raise AugurError if there are multiple
-        # this doesn't matter.
-        # TODO: Remove warning suppression after it's addressed upstream:
-        # <https://github.com/chapmanb/bcbb/issues/143>
-        import warnings
-        from Bio import BiopythonDeprecationWarning
-        warnings.simplefilter("ignore", BiopythonDeprecationWarning)
-        gff_entries = list(GFF.parse(in_handle, limit_info={'gff_type': valid_types}))
-        warnings.simplefilter("default", BiopythonDeprecationWarning)
-
-        if len(gff_entries) == 0:
-            raise AugurError(f"Reference {reference!r} contains no valid data rows. Valid GFF types (3rd column) are {', '.join(valid_types)}.")
-        elif len(gff_entries) > 1:
-            raise AugurError(f"Reference {reference!r} contains multiple seqids (first column). Augur can only handle GFF files with a single seqid.")
-        else:
-            rec = gff_entries[0]
-
-        features['nuc'] = _read_nuc_annotation_from_gff(rec, reference)
-        features_skipped = 0
-
-        for feat in rec.features:
-            if feat.type == "gene":
-                # Check for gene names stored in qualifiers commonly used by
-                # virus-specific gene maps first (e.g., 'gene',
-                # 'gene_name'). Then, check for qualifiers used by non-viral
-                # pathogens (e.g., 'locus_tag').
-                if "gene" in feat.qualifiers:
-                    fname = feat.qualifiers["gene"][0]
-                elif "gene_name" in feat.qualifiers:
-                    fname = feat.qualifiers["gene_name"][0]
-                elif "locus_tag" in feat.qualifiers:
-                    fname = feat.qualifiers["locus_tag"][0]
-                else:
-                    features_skipped+=1
-                    fname = None
-
-                if fname == 'nuc':
-                    raise AugurError(f"Reference {reference!r} contains a gene with the name 'nuc'. This is not allowed.")
-
-                if feature_names is not None and fname not in feature_names:
-                    # Skip (don't store) this feature
-                    continue
-
-                if fname:
-                    features[fname] = feat
-
-        if feature_names is not None:
-            for fe in feature_names:
-                if fe not in features:
-                    print("Couldn't find gene {} in GFF or GenBank file".format(fe))
-
-        if features_skipped:
-            print(f"WARNING: {features_skipped} GFF rows of type=gene skipped as they didn't have a gene, gene_name or locus_tag attribute.")
-
-    return features
-
-def _read_nuc_annotation_from_genbank(record, reference):
-    """
-    Extracts the mandatory 'source' feature. If the sequence is present we check
-    the length agrees with the source. (The 'ORIGIN' may be left blank,
-    according to <https://www.ncbi.nlm.nih.gov/Sitemap/samplerecord.html>.)
-
-    See <https://www.insdc.org/submitting-standards/feature-table/> for more.
-
-    Parameters
-    ----------
-    record : :py:class:`Bio.SeqRecord.SeqRecord` reference: string
-
-    Returns
-    -------
-    :py:class:`Bio.SeqFeature.SeqFeature`
-
-    Raises
-    ------
-    AugurError
-        If 'source' not defined or if coords contradict.
-    """
-    nuc = None
-    for feat in record.features:
-        if feat.type=='source':
-            nuc = feat
-    if not nuc:
-        raise AugurError(f"Reference {reference!r} did not define the mandatory source feature.")
-    if nuc.location.start!=0: # this is a '1' in the GenBank file
-        raise AugurError(f"Reference {reference!r} source feature did not start at 1.")
-    if record.seq and len(record.seq)!=nuc.location.end:
-        raise AugurError(f"Reference {reference!r} source feature was length {nuc.location.end} but the included sequence was length {len(record.seq)}.")
-    return nuc
-
-def _read_genbank(reference, feature_names):
-    """
-    Read a GenBank file. We only read GenBank feature keys 'CDS' or 'source'.
-    We create a "feature name" via:
-    - for 'source' features use 'nuc'
-    - for 'CDS' features use the locus_tag or the gene. If neither, then silently ignore.
-
-    Parameters
-    ----------
-    reference : string
-        File path to GenBank reference
-    feature_names : None or set or list
-        Restrict the CDSs we read to those in the set/list
-
-    Returns
-    -------
-    features : dict
-        keys: feature names, values: :py:class:`Bio.SeqFeature.SeqFeature`
-        Note that feature names may not equivalent to GenBank feature keys
-
-    Raises
-    ------
-    AugurError
-        If 'nuc' annotation not parsed
-        If a CDS feature is given the name 'nuc'
-    """
-    gb = read_single_sequence(reference, format='genbank')
-    features = {
-        'nuc': _read_nuc_annotation_from_genbank(gb, reference)
-    }
-
-    features_skipped = 0
-    for feat in gb.features:
-        if feat.type=='CDS':
-            fname = None
-            if "locus_tag" in feat.qualifiers:
-                fname = feat.qualifiers["locus_tag"][0]
-            elif "gene" in feat.qualifiers:
-                fname = feat.qualifiers["gene"][0]
-            else:
-                features_skipped+=1
-
-            if fname == 'nuc':
-                raise AugurError(f"Reference {reference!r} contains a CDS with the name 'nuc'. This is not allowed.")
-
-            if fname and (feature_names is None or fname in feature_names):
-                features[fname] = feat
-
-    if features_skipped:
-        print(f"WARNING: {features_skipped} CDS features skipped as they didn't have a locus_tag or gene qualifier.")
-
-    return features
 
 def read_lat_longs(overrides=None, use_defaults=True):
     coordinates = {}
